requirements.txt
appnope==0.1.3
asttokens==2.4.0
backcall==0.2.0
bcrypt==4.0.1
blinker==1.6.2
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
decorator==5.1.1
dnspython==2.4.2
email-validator==2.0.0.post2
executing==1.2.0
Flask==2.3.3
Flask-Bcrypt==1.0.1
Flask-DebugToolbar @ git+https://github.com/pallets-eco/flask-debugtoolbar@ce02d2da3c37b3212eebafa9392c429b91d13923
Flask-SQLAlchemy==3.0.5
Flask-WTF==1.1.1
idna==3.4
ipython==8.15.0
itsdangerous==2.1.2
jedi==0.19.0
Jinja2==3.1.2
MarkupSafe==2.1.3
matplotlib-inline==0.1.6
packaging==23.1
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.39
psycopg2-binary==2.9.7
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.16.1
python-dotenv==1.0.0
requests==2.31.0
six==1.16.0
SQLAlchemy==2.0.20
stack-data==0.6.2
traitlets==5.9.0
typing_extensions==4.7.1
urllib3==2.0.4
wcwidth==0.2.6
Werkzeug==2.3.7
WTForms==3.0.1
